完成出入库管理

+++ CMS/ViewModel/CargoViewModel.cs
@@ -1,4 +1,6 @@
-﻿using CMS.Windows;
+﻿using CMS.Entities;
+using CMS.Windows;
+using ControlzEx.Standard;
 using GalaSoft.MvvmLight;
 using GalaSoft.MvvmLight.Command;
 using MahApps.Metro.Controls.Dialogs;
@@ -21,17 +23,47 @@ namespace CMS.ViewModel
             get { return cargoTypes; }
             set { cargoTypes = value; RaisePropertyChanged(); }
         }
-        private List<Cargo> cargos;
+        private List<CargoEx> cargos;
 
-        public List<Cargo> Cargos
+        public List<CargoEx> Cargos
         {
             get { return cargos; }
             set { cargos = value; RaisePropertyChanged(); }
         }
+        public List<CargoEx> SelectCargos()
+        {
+            var cargoes = new CargoProvider().Select();
+            var CargoExs = new List<CargoEx>();
+            foreach (var c in cargoes)
+            {
+                var cargoex = new CargoEx()
+                {
+                    Id = c.Id,
+                    Name = c.Name,
+                    TypeId = c.TypeId,
+                    TypeName = c.TypeName,
+                    Unit = c.Unit,
+                    Price = c.Price,
+                    Tag = c.Tag,
+                    InsertDate = c.InsertDate,
+                    MemberId = c.MemberId,
+                };
+                double sum = 0f;
+                var records = new RecordProvider().Select().FindAll(x => x.CargoName == c.Name);
+                if (records != null)
+                {
+                    var inputcount = records.FindAll(x => x.RecordType == true).Sum(x => x.Number);
+                    var ouputcount = records.FindAll(x => x.RecordType == false).Sum(x => x.Number);
+                    sum=inputcount - ouputcount;
+                }
+                cargoex.Sum = sum;
+                CargoExs.Add(cargoex);
+            }
+            return CargoExs;
+        }
         public CargoViewModel()
         {
-            cargos = new CargoProvider().Select();
-            cargoTypes=new CargoTypeProvider().Select();
+            cargoTypes =new CargoTypeProvider().Select();
         }
 
         public RelayCommand OpenAddCargoWindow
@@ -42,7 +74,7 @@ namespace CMS.ViewModel
                 {
                     var window = new AddCargoWindow();
                     window.ShowDialog();
-                    Cargos = new CargoProvider().Select();
+                    Cargos = SelectCargos();
                 });
             }
         }
@@ -57,10 +89,16 @@ namespace CMS.ViewModel
                         if (MessageBox.Show("确认删除?", "询问", MessageBoxButton.OKCancel, MessageBoxImage.Question)==MessageBoxResult.OK)
                         {
                             CargoProvider cargoProvider = new CargoProvider();
+                            var i = Cargos.FirstOrDefault(item => item.Name == cargo.Name).Sum;
+                            if (i > 0)
+                            {
+                                MessageBox.Show($"当前物资存在库存,无法删除");
+                                return;
+                            }
                             if (cargoProvider.Delete(cargo) > 0)
                             {
                                 AppData.Instance.MainWindow.ShowMessageAsync("提示", "已删除");
-                                Cargos = cargoProvider.Select();
+                                Cargos = SelectCargos();
                             }
                         }
                     }
@@ -83,6 +121,16 @@ namespace CMS.ViewModel
                     }
                 });
 
+            }
+        }
+        public RelayCommand LoadedCommand
+        {
+            get
+            {
+                return new RelayCommand(() =>
+                {
+                    Cargos = SelectCargos();
+                });
             }
         }
     }
